requirements.txt
django==1.6.5
django-mongokit==0.2.6
django-registration==1.0
# django-ratings==0.3.7
django-pagination==1.0.7
django-notification==1.2.0
python-magic==0.4.6
Pillow==2.6.1
django-passwords==0.2.0
# South==0.8.4
# django-extensions==1.3.3
# django-reversion==1.8.0
# gunicorn==19.0.0
# markdown==2.4.1
django-json-rpc==0.6.3
https://bitbucket.org/zalew/django-online-status/get/tip.zip
git+https://github.com/anujag/django-registration-email.git
stemming==1.0.1
numpy==1.9.0
python-memcached==1.58
django-memcache-admin
# pandora-client==0.2.197
django-simple-captcha==0.4.5
hashfs==0.6.0
ox>=2.1.651
celery==3.1.19
django-celery==3.1.16
django-mailbox==4.4.1
Fabric==1.12.0
pathlib2==2.2.1
# pycrypto may need sometimes for Fabric
# pycrypto
sphinx==1.6.5
recommonmark
djangorestframework==3.2.5

# QBANK-LITE REQ: https://github.com/CLIxIndia-Dev/qbank-lite/blob/master/requirements.txt
# also make sure following are updated w.r.t qbank repo
dlkit==0.5.13
beautifulsoup4==4.4.1
cffi==1.7.0
enum34==1.1.6
html5lib==0.9999999
idna==2.1
inflection==0.3.1
ipaddress==1.0.16
lxml==3.6.0
mpmath==0.19
nodeenv==1.0.0
nose==1.3.7
Paste==2.0.3
pyasn1==0.1.9
pycparser==2.14
PyInstaller==3.2.1
pyOpenSSL==16.0.0
signit==0.3.0
six==1.10.0
sympy==1.0
waitress==0.8.10
-e git://github.com/CLIxIndia-Dev/webpy.git@362dcecae9416eef6622922264e6c6482992c898#egg=web.py
bs4==0.0.1
html==1.16
checksumdir==1.1.0
cryptography==1.4
# version needed to overwrite on all above intermediate installs
pymongo==2.8
